% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-md.R
\name{as_chunk_md}
\alias{as_chunk_md}
\title{Markdown chunk}
\usage{
as_chunk_md(x, .from = "markdown", auto_color_link = "blue")
}
\arguments{
\item{x}{A character vector.}

\item{.from}{Pandoc's \code{--from} argument (default: \code{'markdown+autolink_bare_uris+emoji'}).}

\item{auto_color_link}{A color of the link texts.}
}
\description{
Parse markdown cells and returns the "chunk" object.
}
\examples{
library(flextable)
ft <- flextable(data.frame(x = c('**foo**', '**bar**')))
ft <- compose(ft, j = "x", i = 2, value = as_paragraph(as_chunk_md(x)))
autofit(ft)

}
