% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-md.R
\name{parse_md}
\alias{parse_md}
\title{Parse markdown cells}
\usage{
parse_md(x, auto_color_link = "blue", .from = "markdown")
}
\arguments{
\item{x}{A character vector.}

\item{auto_color_link}{A color of the link texts.}

\item{.from}{Pandoc's \code{--from} argument (default: \code{'markdown+autolink_bare_uris+emoji'}).}
}
\description{
Parse markdown cells and returns the "paragraph" object.
}
\examples{
library(flextable)
ft <- flextable(data.frame(x = c('**foo**', '**bar**')))
ft <- compose(ft, j = "x", i = 2, value = parse_md(x))
autofit(ft)

}
