% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-transform.R, R/separate-header.R
\name{separate_header}
\alias{separate_header}
\title{Separate the header based on delimiters}
\usage{
separate_header(x, sep = "[_\\\\.]", fill = FALSE, ...)

separate_header(x, sep = "[_\\\\.]", fill = FALSE, ...)
}
\arguments{
\item{x}{A \code{flextable} object`}

\item{sep}{Separator between columns.

If character, is interpreted as a regular expression. The default
value is a regular expression that matches any sequence of
non-alphanumeric values.

If numeric, interpreted as positions to split at. Positive values start
at 1 at the far-left of the string; negative value start at -1 at the
far-right of the string. The length of \code{sep} should be one less than
\code{into}.}

\item{fill}{If \code{sep} is a character vector, this controls what
happens when there are not enough pieces. There are three valid options:
\itemize{
\item "warn" (the default): emit a warning and fill from the right
\item "right": fill with missing values on the right
\item "left": fill with missing values on the left
}}

\item{...}{Passed to \code{theme_fun}}
}
\description{
Separate the header based on delimiters

Span flextable's header
}
