% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-transform.R
\name{span_header}
\alias{span_header}
\title{Span the header based on delimiters}
\usage{
span_header(x, sep = "[_\\\\.]", theme_fun = flextable::theme_booktabs, ...)
}
\arguments{
\item{x}{A \code{flextable} object`}

\item{sep}{Separator between columns.

If character, \code{sep} is interpreted as a regular expression. The default
value is a regular expression that matches any sequence of
non-alphanumeric values.

If numeric, \code{sep} is interpreted as character positions to split at. Positive
values start at 1 at the far-left of the string; negative value start at -1 at
the far-right of the string. The length of \code{sep} should be one less than
\code{into}.}

\item{theme_fun}{deprecated, use \code{\link[flextable:set_flextable_defaults]{set_flextable_defaults()}} instead.}

\item{...}{Passed to \code{theme_fun}}
}
\description{
Span the header based on delimiters
}
\examples{
iris \%>\%
  as_flextable() \%>\%
  span_header()
}
