% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-keys.R
\name{with_blank}
\alias{with_blank}
\title{Specify blank columns easily via \code{col_keys}}
\usage{
with_blank(after = NULL, before = NULL)
}
\arguments{
\item{after, before}{Blank columns are added after/before the selected columns.
Selections can be done by the semantics of \code{dplyr::select}.}
}
\description{
Specify blank columns easily via \code{col_keys}
}
\examples{
iris \%>\%
  as_flextable(col_keys = with_blank(dplyr::ends_with('Width')))

}
