% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat_md}
\alias{colformat_md}
\title{Format character columns as markdown text}
\usage{
colformat_md(x, auto_color_link = "blue")
}
\arguments{
\item{x}{A \code{flextable} object}

\item{auto_color_link}{A color of the link texts.}
}
\description{
Format character columns as markdown text
}
\examples{
if (rmarkdown::pandoc_available()) {
  data.frame(
    x = c("**bold**", "*italic*" ),
    y = c("^superscript^", "~subscript~"),
    z = c("***^ft^~Extra~** is*", "*Cool*")
  )
}

}
