% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat_md}
\alias{colformat_md}
\title{Format character columns as markdown text}
\usage{
colformat_md(
  x,
  j = where(is.character),
  part = c("body", "header", "all"),
  auto_color_link = "blue",
  .footnote_options = footnote_options(),
  .from = "markdown+autolink_bare_uris",
  .extensions = NULL,
  .pandoc_args = NULL
)
}
\arguments{
\item{x}{A \code{flextable} object}

\item{j}{Columns to be treated as markdown texts.
Selection can be done by the semantics of \code{dplyr::select()}.}

\item{part}{One of "body", "header", and "all". If "all", formatting proceeds in the
order of "header" and "body".}

\item{auto_color_link}{A color of the link texts.}

\item{.footnote_options}{Options for footnotes generated by \code{footnote_options()}.}

\item{.from}{Pandoc's \code{--from} argument (default: \code{'markdown+autolink_bare_uris'}).}

\item{.extensions}{Pandoc's extensions (see \url{https://www.pandoc.org/MANUAL.html#extensions})}

\item{.pandoc_args}{Extra arguments given to Pandoc}
}
\description{
Format character columns as markdown text
}
\examples{
if (rmarkdown::pandoc_available()) {
  d <- data.frame(
    x = c("**bold**", "*italic*"),
    y = c("^superscript^", "~subscript~"),
    z = c("***^ft^~Extra~** is*", "*Cool*")
  )
  colformat_md(flextable::flextable(d))
}
}
