% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat_md}
\alias{colformat_md}
\title{Format character columns as markdown text}
\usage{
colformat_md(
  x,
  j = where(is.character),
  auto_color_link = "blue",
  footnote_ref = c("1", "a", "A", "i", "I", "*"),
  footnote_max = 26,
  footnote_inline = FALSE,
  footnote_sep = "; ",
  .from = "markdown+autolink_bare_uris"
)
}
\arguments{
\item{x}{A \code{flextable} object}

\item{j}{Columns to be treated as markdown texts.
Selection can be done by the semantics of \code{dplyr::select()}.}

\item{auto_color_link}{A color of the link texts.}

\item{footnote_ref}{One of "1", "a", "A", "i", "I", or "*" to as a choice for a symbol to
cross-reference footnotes.}

\item{footnote_max}{A max number of footnotes.}

\item{footnote_inline}{Whether to append footnotes on the same line (default: \code{FALSE}).}

\item{footnote_sep}{A separator of footnotes when \code{footnote_inline = TRUE} (default: \code{"; "}).}

\item{.from}{Pandoc's \code{--from} argument (default: \code{'markdown+autolink_bare_uris'}).}
}
\description{
Format character columns as markdown text
}
\examples{
if (rmarkdown::pandoc_available()) {
  d <- data.frame(
    x = c("**bold**", "*italic*"),
    y = c("^superscript^", "~subscript~"),
    z = c("***^ft^~Extra~** is*", "*Cool*")
  )
  colformat_md(flextable::flextable(d))
}
}
