% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat_md}
\alias{colformat_md}
\title{Format character columns as markdown text}
\usage{
colformat_md(x)
}
\arguments{
\item{x}{A \code{flextable} object}
}
\description{
Format character columns as markdown text
}
\examples{
if (rmarkdown::pandoc_available()) {
  data.frame(
    x = c("**bold**", "*italic*" ),
    y = c("^superscript^", "~subscript~"),
    z = c("***^ft^~Extra~** is*", "*Cool*")
  )
}

}
