% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnote.R
\name{footnote_options}
\alias{footnote_options}
\title{Options for footnotes}
\usage{
footnote_options(
  ref = c("1", "a", "A", "i", "I", "*"),
  prefix = "",
  suffix = "",
  start = 1L,
  max = 26L,
  inline = FALSE,
  sep = "; "
)
}
\arguments{
\item{ref}{One of "1", "a", "A", "i", "I", or "*" to as a choice for a symbol to
cross-reference footnotes.}

\item{prefix, suffix}{Pre- and suf-fixes for \code{ref} (default: \code{""}).}

\item{start}{A starting number of footnotes.}

\item{max}{A max number of footnotes.}

\item{inline}{whether to add footnote on same line as previous footnote or not}

\item{sep}{used only when inline = TRUE, character string to use as
a separator between footnotes.}
}
\value{
An environment
}
\description{
Configure options for footnotes.
}
\examples{
o <- footnote_options("1", start = 1L)

}
